// backend/supabase/functions/_shared/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
      order: {
        Row: {
          adsFee: number | null
          amountBuy: string
          amountBuyFilled: string | null
          amountSell: string
          createdAt: string | null
          deadline: string | null
          desiredTaker: string | null
          id: string
          maker: string
          nonce: string
          signature: string
          status: Database["public"]["Enums"]["order_status"] | null
          tokenBuy: string
          tokenSell: string
        }
        Insert: {
          adsFee?: number | null
          amountBuy: string
          amountBuyFilled?: string | null
          amountSell: string
          createdAt?: string | null
          deadline?: string | null
          desiredTaker?: string | null
          id?: string
          maker: string
          nonce: string
          signature: string
          status?: Database["public"]["Enums"]["order_status"] | null
          tokenBuy: string
          tokenSell: string
        }
        Update: {
          adsFee?: number | null
          amountBuy?: string
          amountBuyFilled?: string | null
          amountSell?: string
          createdAt?: string | null
          deadline?: string | null
          desiredTaker?: string | null
          id?: string
          maker?: string
          nonce?: string
          signature?: string
          status?: Database["public"]["Enums"]["order_status"] | null
          tokenBuy?: string
          tokenSell?: string
        }
        Relationships: [
          {
            foreignKeyName: "order_tokenBuy_fkey"
            columns: ["tokenBuy"]
            isOneToOne: false
            referencedRelation: "token"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "order_tokenSell_fkey"
            columns: ["tokenSell"]
            isOneToOne: false
            referencedRelation: "token"
            referencedColumns: ["id"]
          },
        ]
      }
      token: {
        Row: {
          decimals: number
          id: string
          name: string
          symbol: string
        }
        Insert: {
          decimals: number
          id: string
          name: string
          symbol: string
        }
        Update: {
          decimals?: number
          id?: string
          name?: string
          symbol?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      order_status: "canceled" | "filled" | "active"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never